+++ CHANGELOG.md
@@ -7,6 +7,10 @@ Adheres to [Semantic Versioning](http://semver.org/).
 ## 1.1.1 (TBD)
 
 * Java 11
+* ETRS89 GeoDatums enumeration
+* Additional included names in Ellipsoids
+* Ellipsoids saved reciprocal flattening value
+* Prime Meridians name capitalizations
 
 ## [1.1.0](https://github.com/ngageoint/coordinate-reference-systems-java/releases/tag/1.1.0) (10-14-2021)